v3.7.0 =====================================================================
--- New Features -------------------------

- Github Actions integration from 3.7.0 and up!

```yaml
  # try it out in gh actions
  - uses: wrynegade/scwrypts@main
    with:
      version: v3.7.0
      scwrypt: --name hello-world --group scwrypts --type py
      args: --message "hello from github actions ci <3"
```

--- New Scripts --------------------------

zsh/helm )
  smart helm template functions (simply pass a filename)
   - get-template
   - update-dependencies

--- Changes ------------------------------

- CHECK_ENVIRONMENT now uses proper argument parsing

- scwrypts/plugins loaded by setting in config or environment:
   SCWRYPTS_PLUGIN_ENABLED__plugin=1

- SCWRYPTS__GET_PATH_TO_RELATIVE_ARGUMENT was missed in the v2->v3
  refactor and has now been reincluded as SCWRYPTS__GET_REALPATH

+++ zsh/lib/scwrypts/virtualenv.module.zsh
@@ -1,9 +1,6 @@
 #####################################################################
 
-DEPENDENCIES+=(
-	virtualenv
-	nodeenv
-)
+DEPENDENCIES+=()
 REQUIRED_ENV+=()
 
 use utils
@@ -15,14 +12,14 @@ AVAILABLE_VIRTUALENVS=(py zx)
 REFRESH_VIRTUALENV() {
 	local GROUP="$1"
 	local TYPE="$2"
-	[ ! $TYPE ] && {
-		ERROR 'no virtualenv type specified'
-		return 1
-	}
-	STATUS "refreshing $GROUP/$TYPE virtual environment"
+	local VIRTUALENV_PATH="$(_VIRTUALENV__GET_PATH)"
+
+	[ ! $TYPE ] && { ERROR 'no virtualenv type specified'; return 1; }
+
+	STATUS "refreshing $GROUP/$TYPE virtualenv"
 	DELETE_VIRTUALENV $GROUP $TYPE \
 		&& UPDATE_VIRTUALENV $GROUP $TYPE \
-		&& SUCCESS 'successfully refreshed virtual environment' \
+		&& SUCCESS 'successfully refreshed virtualenv' \
 		|| { ERROR 'something went wrong during refresh (see above)'; return 1; } \
 		;
 }
@@ -30,64 +27,50 @@ REFRESH_VIRTUALENV() {
 UPDATE_VIRTUALENV() {
 	local GROUP="$1"
 	local TYPE="$2"
-	[ ! $TYPE ] && {
-		ERROR 'no virtualenv type specified'
-		return 1
-	}
+	local VIRTUALENV_PATH="$(_VIRTUALENV__GET_PATH)"
 
-	local VIRTUALENV_PATH=$(GET_VIRTUALENV_PATH $GROUP $TYPE)
+	[ ! $TYPE ] && { ERROR 'no virtualenv type specified'; return 1; }
 
-	[ ! -d $VIRTUALENV_PATH ] && {
-		which CREATE_VIRTUALENV__${GROUP}__${TYPE} >/dev/null 2>&1 \
-			&& CREATE_VIRTUALENV__${GROUP}__${TYPE} $VIRTUALENV_PATH \
-			|| return 0
-	}
+	: \
+		&& which CREATE_VIRTUALENV__${GROUP}__${TYPE} >/dev/null 2>&1 \
+		&& which ACTIVATE_VIRTUALENV__${GROUP}__${TYPE} >/dev/null 2>&1 \
+		&& which UPDATE_VIRTUALENV__${GROUP}__${TYPE} >/dev/null 2>&1 \
+		&& which DEACTIVATE_VIRTUALENV__${GROUP}__${TYPE} >/dev/null 2>&1 \
+		|| { STATUS "no virtualenv available for $GROUP/$TYPE; skipping"; return 0; }
 
-	STATUS "updating $TYPE virtual environment"
-
-	source $VIRTUALENV_PATH/bin/activate || {
-		ERROR 'failed to activate virtualenv; did create fail?'
-		return 1
-	}
-
-	cd $SCWRYPTS_ROOT
-	local UPDATE_CODE=0
-	case $TYPE in
-		py ) cd py; pip install --no-cache-dir -r requirements.txt; UPDATE_CODE=$? ;;
-		zx ) cd zx; npm install ;;
-	esac
-	UPDATE_CODE=$?
-	[[ $UPDATE_CODE -eq 0 ]] \
-		&& SUCCESS "$TYPE virtual environment up-to-date" \
-		|| ERROR "failed to update $TYPE virtual environment (see errors above)" \
-		;
-
-	deactivate_node >/dev/null 2>&1
-	deactivate >/dev/null 2>&1
-	return $UPDATE_CODE
+	STATUS "updating $GROUP/$TYPE virtualenv" \
+		&& CREATE_VIRTUALENV__${GROUP}__${TYPE} \
+		&& ACTIVATE_VIRTUALENV__${GROUP}__${TYPE} \
+		&& UPDATE_VIRTUALENV__${GROUP}__${TYPE} \
+		&& DEACTIVATE_VIRTUALENV__${GROUP}__${TYPE} \
+		&& SUCCESS "$GROUP/$TYPE virtualenv up-to-date" \
+		|| { ERROR "failed to update $GROUP/$TYPE virtualenv (see errors above)"; return 2; }
 }
 
 DELETE_VIRTUALENV() {
+	[ $CI ] && return 0
+
 	local GROUP="$1"
 	local TYPE="$2"
-	local VIRTUALENV_PATH="$(GET_VIRTUALENV_PATH $GROUP $TYPE)"
+	local VIRTUALENV_PATH="$(_VIRTUALENV__GET_PATH)"
 
-	STATUS "dropping $TYPE virtual environment artifacts"
+	[ ! $TYPE ] && { ERROR 'no virtualenv type specified'; return 1; }
+
+	STATUS "dropping $GROUP/$TYPE virtualenv artifacts"
 
 	[ ! -d $VIRTUALENV_PATH ] && {
-		SUCCESS "no $TYPE environment detected"
+		SUCCESS "no $GROUP/$TYPE environment detected"
 		return 0
 	}
 
 	rm -rf $VIRTUALENV_PATH \
-		&& SUCCESS "succesfully cleaned up $TYPE virtual environment" \
+		&& SUCCESS "succesfully cleaned up $GROUP/$TYPE virtualenv" \
 		|| { ERROR "unabled to remove '$VIRTUALENV_PATH'"; return 1; }
 }
 
-GET_VIRTUALENV_PATH() {
-	local GROUP="$1"
-	local TYPE="$2"
+#####################################################################
 
+_VIRTUALENV__GET_PATH() {
 	local ENV_PATH="$(eval echo '$SCWRYPTS_VIRTUALENV_PATH__'$GROUP 2>/dev/null)"
 	[ ! $ENV_PATH ] && ENV_PATH="$SCWRYPTS_VIRTUALENV_PATH__scwrypts"
 
@@ -97,9 +80,14 @@ GET_VIRTUALENV_PATH() {
 #####################################################################
 
 CREATE_VIRTUALENV__scwrypts__py() {
+	[ $CI ] && return 0
+	[ -d $VIRTUALENV_PATH ] && return 0
+
+	DEPENDENCIES=(virtualenv) CHECK_ENVIRONMENT || return 1
+
 	local VIRTUALENV_PATH="$1"
 
-	STATUS 'creating python virtual environment'
+	STATUS 'creating python virtualenv'
 	local PY PYTHON
 	for PY in $(echo $SCWRYPTS_PREFERRED_PYTHON_VERSIONS__scwrypts)
 	do
@@ -122,7 +110,37 @@ CREATE_VIRTUALENV__scwrypts__py() {
 		}
 }
 
+ACTIVATE_VIRTUALENV__scwrypts__py() {
+	[ $CI ] && return 0
+	source $VIRTUALENV_PATH/bin/activate || {
+		ERROR "failed to activate virtualenv $GROUP/$TYPE; did create fail?"
+		return 1
+	}
+}
+
+UPDATE_VIRTUALENV__scwrypts__py() {
+	local PIP_INSTALL_ARGS=()
+
+	PIP_INSTALL_ARGS+=(--no-cache-dir)
+	PIP_INSTALL_ARGS+=(-r requirements.txt)
+
+	cd "$SCWRYPTS_ROOT/py"
+	pip install ${PIP_INSTALL_ARGS[@]}
+}
+
+DEACTIVATE_VIRTUALENV__scwrypts__py() {
+	deactivate >/dev/null 2>&1
+	return 0
+}
+
+##########################################
+
 CREATE_VIRTUALENV__scwrypts__zx() {
+	[ $CI ] && return 0
+	[ -d $VIRTUALENV_PATH ] && return 0
+
+	DEPENDENCIES=(nodeenv) CHECK_ENVIRONMENT || return 1
+
 	local VIRTUALENV_PATH="$1"
 
 	STATUS 'setting up nodeenv'
@@ -132,4 +150,24 @@ CREATE_VIRTUALENV__scwrypts__zx() {
 			ERROR "unable to create '$VIRTUALENV_PATH' with '$SCWRYPTS_NODE_VERSION__scwrypts'"
 			return 2
 		}
+}
+
+ACTIVATE_VIRTUALENV__scwrypts__zx() {
+	[ $CI ] && return 0
+	source $VIRTUALENV_PATH/bin/activate || {
+		ERROR "failed to activate virtualenv $GROUP/$TYPE; did create fail?"
+		return 1
+	}
+}
+
+UPDATE_VIRTUALENV__scwrypts__zx() {
+	local NPM_INSTALL_ARGS=()
+
+	cd "$SCWRYPTS_ROOT/zx"
+	npm install ${NPM_INSTALL_ARGS[@]}
+}
+
+DEACTIVATE_VIRTUALENV__scwrypts__zx() {
+	deactivate_node >/dev/null 2>&1
+	return 0
 }
